requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
mysql-connector-python==9.0.0
sqlglot==25.16.0
httpx==0.27.0
python-dotenv==1.0.1
